+++ scripts/simplify_network.py
@@ -62,7 +62,7 @@ Outputs
     .. image:: ../img/regions_offshore_elec_s  .png
             :scale: 33 %
 
-- ``resources/clustermaps_{network}_s{simpl}.h5``: Mapping of buses from ``networks/elec.nc`` to ``networks/elec_s{simpl}.nc``; has keys ['/busmap_s']
+- ``resources/busmap_{network}_s{simpl}.csv``: Mapping of buses from ``networks/elec.nc`` to ``networks/elec_s{simpl}.nc``;
 - ``networks/{network}_s{simpl}.nc``:
 
     .. image:: ../img/elec_s.png
@@ -315,7 +315,7 @@ def remove_stubs(n):
     return n, busmap
 
 def cluster(n, n_clusters):
-    logger.info("Clustering to {} buses".format(n_clusters))
+    logger.info(f"Clustering to {n_clusters} buses")
 
     renewable_carriers = pd.Index([tech
                                     for tech in n.generators.carrier.unique()
@@ -329,7 +329,7 @@ def cluster(n, n_clusters):
     potential_mode = (consense(pd.Series([snakemake.config['renewable'][tech]['potential']
                                             for tech in renewable_carriers]))
                         if len(renewable_carriers) > 0 else 'conservative')
-    clustering = clustering_for_n_clusters(n, n_clusters, potential_mode=potential_mode,
+    clustering = clustering_for_n_clusters(n, n_clusters, custom_busmap=False, potential_mode=potential_mode,
                                            solver_name=snakemake.config['solving']['solver']['name'])
 
     return clustering.network, clustering.busmap